refactor: add usePianoRollKeyboardShortcut
